Refactorizada estructura edl proyecto. Agregado Lerna, ESLint y Husky
