+++ apps/backend/index.js
@@ -1,0 +1,6 @@
+const app = require('./src/interfaces/http/server');
+const PORT = process.env.PORT || 3000;
+
+app.listen(PORT, () => {
+  console.log(`Servidor corriendo en el puerto ${PORT}`);
+});
