+++ apps/backend/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "backend",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "start": "node --watch index.js",
+    "test": "jest",
+    "test:coverage": "jest --coverage"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "",
+  "type": "commonjs",
+  "dependencies": {
+    "cors": "2.8.5",
+    "express": "5.1.0",
+    "dotenv": "17.2.1",
+    "pg": "8.16.3"
+  },
+  "devDependencies": {
+    "jest": "30.1.3",
+    "supertest": "7.1.4"
+  }
+}
