+++ apps/backend/src/domain/escaladores/Escalador.js
@@ -1,0 +1,10 @@
+class Escalador {
+  constructor({ id, nombre, edad, experiencia }) {
+    this.id = id;
+    this.nombre = nombre;
+    this.edad = edad;
+    this.experiencia = experiencia;
+  }
+}
+
+module.exports = Escalador;
