+++ apps/backend/src/interfaces/http/controllers/pistaController.js
