+++ apps/backend/src/interfaces/http/routes/pistas.routes.js
