+++ apps/frontend/index.html
@@ -1,0 +1,17 @@
+<!doctype html>
+<html lang="es">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Mi APP</title>
+  </head>
+
+  <body>
+    <h1>Prueba de conexión</h1>
+
+    <button id="test-api">Llamar a la API</button>
+    <pre id="api-response"></pre>
+
+    <script src="js/main.js"></script>
+  </body>
+</html>
