+++ apps/frontend/views/escaladores/formulario.html
@@ -1,0 +1,47 @@
+<!doctype html>
+<html lang="es">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>Crear Escalador</title>
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+    />
+  </head>
+  <body class="container py-5">
+    <h1>Nuevo Escalador</h1>
+    <form action="http://localhost:3000/escaladores" method="POST">
+      <div class="mb-3">
+        <label for="nombre" class="form-label">Nombre</label>
+        <input
+          type="text"
+          class="form-control"
+          name="nombre"
+          id="nombre"
+          required
+        />
+      </div>
+      <div class="mb-3">
+        <label for="edad" class="form-label">Edad</label>
+        <input
+          type="number"
+          class="form-control"
+          name="edad"
+          id="edad"
+          required
+        />
+      </div>
+      <div class="mb-3">
+        <label for="experiencia" class="form-label">Experiencia</label>
+        <textarea
+          class="form-control"
+          name="experiencia"
+          id="experiencia"
+          rows="3"
+        ></textarea>
+      </div>
+      <button type="submit" class="btn btn-primary">Crear</button>
+    </form>
+  </body>
+</html>
