--- index.js
@@ -1,6 +1,0 @@
-const app = require('./src/interfaces/http/server');
-const PORT = process.env.PORT || 3000;
-
-app.listen(PORT, () => {
-  console.log(`Servidor corriendo en puerto ${PORT}`);
-});
