+++ package.json
@@ -1,25 +1,40 @@
 {
-  "name": "climbit",
+  "name": "prueba-monorepo",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
+  "private": true,
+  "workspaces": [
+    "apps/*"
+  ],
   "scripts": {
-    "test": "jest",
-    "test:coverage": "jest --coverage",
-    "dev": "nodemon index.js"
+    "dev": "lerna run start --parallel",
+    "test": "lerna run test",
+    "test:coverage": "lerna run test:coverage",
+    "clean": "lerna clean",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix",
+    "format": "prettier --write .",
+    "prepare": "husky"
   },
   "keywords": [],
   "author": "",
-  "license": "ISC",
+  "license": "",
   "type": "commonjs",
-  "dependencies": {
-    "dotenv": "^17.2.1",
-    "express": "^5.1.0",
-    "pg": "^8.16.3"
-  },
   "devDependencies": {
-    "jest": "^30.1.3",
-    "nodemon": "^3.1.10",
-    "supertest": "^7.1.4"
+    "@eslint/js": "9.37.0",
+    "eslint": "9.37.0",
+    "eslint-config-google": "0.14.0",
+    "eslint-config-prettier": "10.1.8",
+    "eslint-plugin-jest": "29.0.1",
+    "eslint-plugin-prettier": "5.5.4",
+    "globals": "16.4.0",
+    "husky": "9.1.7",
+    "lerna": "9.0.0",
+    "lint-staged": "16.2.4",
+    "prettier": "3.6.2"
+  },
+  "lint-staged": {
+    "**/*.{js,mjs}": "eslint --fix"
   }
 }
