--- src/domain/escaladores/Escalador.js
@@ -1,11 +1,0 @@
-class Escalador {
-  constructor({ id, nombre, edad, experiencia }) {
-    this.id = id;
-    this.nombre = nombre;
-    this.edad = edad;
-    this.experiencia = experiencia;
-  }
-}
-
-module.exports = Escalador;
-
