--- src/interfaces/http/controllers/escaladorController.js
@@ -1,14 +1,0 @@
-const crearEscaladorUseCase = require('../../../application/escaladores/crearEscalador');
-const repository = require('../../../infrastructure/repositories/EscaladorRepositoryPostgres');
-
-const crearEscalador = async (req, res) => {
-  try {
-    const crear = crearEscaladorUseCase(repository);
-    const nuevo = await crear(req.body);
-    res.status(201).json(nuevo);
-  } catch (err) {
-    res.status(500).json({ error: err.message });
-  }
-};
-
-module.exports = { crearEscalador };
