--- src/interfaces/web/views/pistas/formulario.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <title>Crear Pista</title>
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" />
-</head>
-<body class="container py-5">
-  <h1>Nueva Pista</h1>
-  <form action="/pistas" method="POST">
-    <div class="mb-3">
-      <label for="nombre" class="form-label">Nombre</label>
-      <input type="text" class="form-control" name="nombre" id="nombre" required />
-    </div>
-    <div class="mb-3">
-      <label for="dificultad" class="form-label">Dificultad</label>
-      <select class="form-control" name="dificultad" id="dificultad">
-        <option value="Fácil">Fácil</option>
-        <option value="Media">Media</option>
-        <option value="Difícil">Difícil</option>
-      </select>
-    </div>
-    <button type="submit" class="btn btn-primary">Crear</button>
-  </form>
-</body>
-</html>
